Split Fixed Points into two implementations.
